style: make the code easier to maintain
hear me out: prefix your functions with f.
this is more efficient, especially if you have an editor/ide that
supports autocompletion; if you wanna call a function, just press `f`
and the list of all the functions show up in one place.

similarly to hungarian notation; if you're expecting a `bool`, then
press `b` to pull up all the booleans in the current scope. highly
efficient once you get used to it, and also easy to read (you dont have
to move your mouse or see where the variable is defined to know the type)

this is not a hot take. i know i am correct and those who disagree
probably use the mouse to change lines